+++ src/entities/auth/model/authorized-user.store.ts
@@ -1,12 +1,8 @@
-import { createEvent, createStore } from 'effector';
+import { restore } from 'effector';
 
-import { type OperationInfo } from '@specialized-messenger/api/specs';
+import { baseSignInFx } from './base-sign-in.effect';
 
-type User = OperationInfo<'AuthController_signIn_v1'>['response'];
-
-export const $authorizedUser = createStore<User | null>(null);
+export const $authorizedUser = restore(baseSignInFx.doneData, null).map(
+  (data) => data?.user || null,
+);
 export const $isAuthorized = $authorizedUser.map(Boolean);
-
-export const setAuthorizedUser = createEvent<User | null>();
-
-$authorizedUser.on(setAuthorizedUser, (_, payload) => payload);