+++ src/entities/auth/model/base-sign-in.effect.ts
@@ -4,9 +4,7 @@ import type { OperationInfo } from '@specialized-messenger/api/specs';
 
 import { api } from '@/shared/api';
 import { saveCredentialsInLocalStorage } from '@/shared/lib/auth';
-import { showErrorNotificationFx } from '@/shared/lib/notifications';
-
-import { setAuthorizedUser } from '@/entities/auth';
+import type { Credentials } from '@/shared/lib/auth/credentials';
 
 type Controller = OperationInfo<'AuthController_signIn_v1'>;
 type Path = Controller['path'];
@@ -17,23 +15,21 @@ export const baseSignInFx = createEffect<
   {
     requestBody: Body;
   },
-  void
->(async ({ requestBody }) => {
-  try {
-    const user = await api
-      .post<Response>('/api/v1/auth/sign-in' satisfies Path, {
-        json: requestBody,
-      })
-      .json();
-
-    saveCredentialsInLocalStorage({
-      login: requestBody.login,
-      password: requestBody.password,
-    });
-    setAuthorizedUser(user);
-  } catch (error) {
-    showErrorNotificationFx({ message: 'Authorization error' });
-
-    throw error;
+  {
+    user: Response;
+    credentials: Credentials;
   }
+>(async ({ requestBody }) => {
+  const user = await api
+    .post<Response>('/api/v1/auth/sign-in' satisfies Path, {
+      json: requestBody,
+    })
+    .json();
+
+  saveCredentialsInLocalStorage(requestBody);
+
+  return {
+    user,
+    credentials: requestBody,
+  };
 });