+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [1.4.2](https://github.com/RomanDeveloperGit/specialized-messenger-frontend/compare/v1.4.1...v1.4.2) (2026-04-22)
+
+
+### Bug Fixes
+
+* **invitation:** add error plate ([60b59f0](https://github.com/RomanDeveloperGit/specialized-messenger-frontend/commit/60b59f016024a34e6e093e5c138a996bd5cbe8e0))
+
 ## [1.4.1](https://github.com/RomanDeveloperGit/specialized-messenger-frontend/compare/v1.4.0...v1.4.1) (2026-04-21)
 
 ### Bug Fixes